Updated BaseContainer to render each section.
To ensure consistency across all the different sections, the BaseContainer is now designed to be subclassed rather than composed in.  Each of the different sections, Example, Code, and Documentation all have their own render method that is called by the main render method to ensure that everything comes together correctly.  This should remove nearly all the boilerplate code necessary in each component container page.

NOTE: This will require all the current containers to be updated.

In the future, I would like to figure out a way that we can get the sample code and example using the same information so that it automatically updates both when it is changed.  That requires some wizardry I haven’t figured out yet, however, so for the moment, copy and paste it is.